requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
huggingface_hub==0.24.6
httpx==0.27.2
PyYAML==6.0.2
click==8.1.7
